Update projects NuGet references

+++ source/Hackster/McpLedBarGraph/MeadowApp.cs
@@ -52,7 +52,7 @@ namespace McpLedBarGraph
                 Console.WriteLine("Turning them on using SetLed...");
                 for (int i = 0; i < ledBarGraph.Count; i++)
                 {
-                    ledBarGraph.SetLed(i, true);
+                    ledBarGraph.SetLed((uint)i, true);
                     Thread.Sleep(300);
                 }
 
@@ -61,7 +61,7 @@ namespace McpLedBarGraph
                 Console.WriteLine("Turning them off using SetLed...");
                 for (int i = ledBarGraph.Count - 1; i >= 0; i--)
                 {
-                    ledBarGraph.SetLed(i, false);
+                    ledBarGraph.SetLed((uint)i, false);
                     Thread.Sleep(300);
                 }
